Improve Merge Sort Performance
The Performance difference to chrome had mainly two roots (unsafe mode,
safe mode still needs to be optimized)

1. Conditions in if expressions can often be simplified to directly
descend into the if or else branch if the condition is true / false instead
of creating own branches (e.g. for an && or || expression). The old approach
was hindering the optimizer from doing good work

2. Optimizations on pure pointers seem to be not well supported by the optimizier.
But, it is if C++ references (or dereferenceables) are passed instead of
naked pointers. Therefore, mark parameters as dereferenceable if it is
a pointer to an object.

This changes make the unsafe implementation as fast as the safe implementation.
The reason why the safe implementation is only half as fast needs further
investigation that is currently postponed.