Do not use splice in tspLarge
Array.splice has been specially optimized in Chrome to use sparse array
in case the array contains more than 1000 elements and only a small number
is going to be deleted. A sparse array does not actually move all elements,
instead, deleted elements are marked as holes. Therefore, the
deletion operation is nearly of constant time (except when the array
contains too many  holes, in this case the array is compacted).

As it is not the main goal to implement the most efficient array in this
thesis rewrite the tsp large benchmark to not use splice.