Support passing objects
Adds support for passing objects from JS to WASM and from WASM to JS.
The implementation maintains reference equality between arguments.
However, the call semantics differ from normal JavaScript as it is by
value instead of by reference. Meaning, that changes to the JavaScript
objects performed in the WASM functions are not reflected to the
objects in JavaScript.