Refactor Math runtime functions to accept this parameter
Refactor the math methods to accept the this object as first argument.
Otherwise a special handling just for Math objects are needed what adds
non inherent complexity. Therefore, pass this to all math functions
even if it is not needed and let the compiler remove the unnecessary loads
during optimization.