Add Casts for runtime function call arguments and return values if needed
The runtime only knows 4 types:
* bool
* int
* double
* and a pointer

On the contrary, Speedy.js allows to define new object types as well that
are not known to the runtime. This requires, that objects passed
to runtime methods are casted to i8* ptrs and casted back to the actual
type if returned by a runtime function.

Additionally: Added support for min, floor, round